Feat: rename files and update storybook
