+++ package.json
@@ -24,16 +24,13 @@
     "react-dom": "^19.0.0"
   },
   "devDependencies": {
-    "@chromatic-com/storybook": "^3.2.6",
+    "@chromatic-com/storybook": "^4.0.0",
     "@lhci/cli": "^0.14.0",
-    "@storybook/addon-essentials": "^8.6.12",
-    "@storybook/addon-onboarding": "^8.6.12",
-    "@storybook/blocks": "^8.6.12",
-    "@storybook/experimental-addon-test": "^8.6.12",
-    "@storybook/experimental-nextjs-vite": "^8.6.12",
-    "@storybook/nextjs": "^8.6.12",
-    "@storybook/react": "^8.6.12",
-    "@storybook/test": "^8.6.12",
+    "@storybook/addon-docs": "^9.0.1",
+    "@storybook/addon-onboarding": "^9.0.1",
+    "@storybook/addon-vitest": "^9.0.1",
+    "@storybook/nextjs": "^9.0.1",
+    "@storybook/nextjs-vite": "^9.0.1",
     "@tailwindcss/postcss": "^4.1.5",
     "@types/node": "^20",
     "@types/react": "^19",
@@ -44,10 +41,13 @@
     "daisyui": "^5.0.35",
     "playwright": "^1.52.0",
     "start-server-and-test": "^2.0.12",
-    "storybook": "^8.6.12",
+    "storybook": "^9.0.1",
     "tailwindcss": "^4.1.5",
     "typescript": "^5",
     "vitest": "^3.1.2",
     "wait-on": "^8.0.3"
+  },
+  "overrides": {
+    "storybook": "$storybook"
   }
 }
