+++ src/stories/Configure.mdx
@@ -1,4 +1,4 @@
-import { Meta } from "@storybook/blocks";
+import { Meta } from "@storybook/addon-docs/blocks";
 import Image from "next/image";
 
 import Github from "./assets/github.svg";
