+++ src/stories/Header.stories.ts
@@ -1,4 +1,4 @@
-import type { Meta, StoryObj } from '@storybook/react';
+import type { Meta, StoryObj } from '@storybook/nextjs';
 import { Header, HeaderProps } from './Header';
 
 const meta: Meta<typeof Header> = {
