+++ src/stories/Paragraph.stories.ts
@@ -1,5 +1,5 @@
-import type { Meta, StoryObj } from '@storybook/react';
-import { expect, userEvent, within } from '@storybook/test';
+import type { Meta, StoryObj } from '@storybook/nextjs';
+import { expect, userEvent, within } from 'storybook/test';
 
 import { Paragraph } from './Paragraph';
